feat(FE-memos): 스크롤 관련 스타일 추가
수평 스크롤, 스크롤 없을 때 그라데이션
